Создание интерфейсов для сервисов, реализована работа агрегаторов
