+++ telemetry/analyzer/src/main/java/ru/yandex/practicum/aggregator/SnapshotAggregator.java
@@ -10,7 +10,8 @@ import ru.yandex.practicum.kafka.telemetry.event.SensorsSnapshotAvro;
 import ru.yandex.practicum.receiver.BaseAggregator;
 import ru.yandex.practicum.receiver.KafkaConsumerManager;
 import ru.yandex.practicum.receiver.OffsetCommitManager;
-import ru.yandex.practicum.record_process.SnapshotButchProcessor;
+import ru.yandex.practicum.record_process.EventButchProcessor;
+import ru.yandex.practicum.record_process.RecordProcessor;
 
 import java.util.List;
 
@@ -18,11 +19,14 @@ import java.util.List;
 @Component
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SnapshotAggregator extends BaseAggregator<String, SensorsSnapshotAvro> {
-
+final RecordProcessor<SensorsSnapshotAvro> recordProcessor;
     @Autowired
     public SnapshotAggregator(KafkaConsumerManager<String, SensorsSnapshotAvro> consumerManager,
-                              OffsetCommitManager<String, SensorsSnapshotAvro> offsetCommitManager) {
+                              OffsetCommitManager<String, SensorsSnapshotAvro> offsetCommitManager,
+            RecordProcessor<SensorsSnapshotAvro> recordProcessor)
+    {
         super (consumerManager, offsetCommitManager);
+        this.recordProcessor = recordProcessor;
     }
     @Value("${kafka.topics.snapshots_topic}")
     private String inputTopic;
@@ -32,11 +36,12 @@ public class SnapshotAggregator extends BaseAggregator<String, SensorsSnapshotAv
         return List.of(inputTopic);
     }
     @Override
-protected SnapshotButchProcessor<String, SensorsSnapshotAvro> createBatchProcessor()
+protected EventButchProcessor<String, SensorsSnapshotAvro> createBatchProcessor()
     {
-       return new SnapshotButchProcessor<>(
+       return new EventButchProcessor<>(
                offsetCommitManager,
-               processing
+               processing,
+               recordProcessor
        );
     }
 }
