+++ telemetry/analyzer/src/main/java/ru/yandex/practicum/record_process/EventButchProcessor.java
@@ -6,15 +6,14 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import ru.yandex.practicum.receiver.OffsetCommitManager;
 
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
-
 @RequiredArgsConstructor
 @Slf4j
-public class SnapshotButchProcessor<K,V> implements Consumer<ConsumerRecords<K, V>> {
+public class EventButchProcessor<K,V> implements Consumer<ConsumerRecords<K, V>> {
     private final OffsetCommitManager<K, V> offsetCommitManager;
     private final AtomicBoolean processing;
+    private final RecordProcessor<V> recordProcessor;
     @Override
     public void accept(ConsumerRecords<K, V> records) {
         processing.set(true);
@@ -22,13 +21,11 @@ public class SnapshotButchProcessor<K,V> implements Consumer<ConsumerRecords<K, 
             for (ConsumerRecord<K, V> record : records) {
                 try {
                     offsetCommitManager.recordProcessed(record);
-                    log.info("Получена снапшот для обработки {}", record.value());
-                    //обработка одного снапшота - результат в boolean
-                    //если boolean - отправка
-//                    Optional<R> result = recordProcessor.process(record.value());
-//                    result.ifPresent(snapshotHandler::handle);
-                } catch (Exception e) {
-                    log.error("Ошибка обработки снапшота", e);
+                    log.info("Получена запись для обработки {}", record.value());
+                   var event = record.value();
+                    recordProcessor.process(event);
+                   } catch (Exception e) {
+                    log.error("Ошибка обработки записи", e);
                 }
             }
         } finally {
