+++ telemetry/analyzer/src/main/java/ru/yandex/practicum/record_process/SnapshotProcessor.java
@@ -4,11 +4,9 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.jdbc.datasource.UserCredentialsDataSourceAdapter;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.hub_executor.HubActionSender;
 import ru.yandex.practicum.kafka.telemetry.event.ClimateSensorEventAvro;
-import ru.yandex.practicum.kafka.telemetry.event.HubEventAvro;
 import ru.yandex.practicum.kafka.telemetry.event.LightSensorEventAvro;
 import ru.yandex.practicum.kafka.telemetry.event.MotionSensorEventAvro;
 import ru.yandex.practicum.kafka.telemetry.event.SensorStateAvro;
@@ -21,6 +19,7 @@ import ru.yandex.practicum.model.ConditionOperationType;
 import ru.yandex.practicum.model.Scenario;
 import ru.yandex.practicum.model.Sensor;
 import ru.yandex.practicum.service.ScenarioService;
+
 
 import java.util.List;
